ZMS-Conf: Support Check for Updates
