+++ Products/zms/conf/metacmd_manager/manage_css_classes/__init__.py
@@ -13,7 +13,7 @@ class manage_css_classes:
 	description = ""
 
 	# Execution
-	execution = 0
+	execution = False
 
 	# Icon_clazz
 	icon_clazz = "fab fa-css3 text-primary"
