+++ Products/zms/conf/metacmd_manager/manage_repository_gitpush/__init__.py
@@ -13,7 +13,7 @@ class manage_repository_gitpush:
 	description = ""
 
 	# Execution
-	execution = 0
+	execution = False
 
 	# Icon_clazz
 	icon_clazz = "fas fa-forward"
