+++ Products/zms/conf/metacmd_manager/manage_tab_ZMSIndexZCatalog/__init__.py
@@ -13,7 +13,7 @@ class manage_tab_ZMSIndexZCatalog:
 	description = "ZMSIndex"
 
 	# Execution
-	execution = 0
+	execution = False
 
 	# Icon_clazz
 	icon_clazz = "fas fa-cogs"
