+++ Products/zms/conf/metacmd_manager/manage_translate/__init__.py
@@ -13,7 +13,7 @@ class manage_translate:
 	description = ""
 
 	# Execution
-	execution = 0
+	execution = False
 
 	# Icon_clazz
 	icon_clazz = "fas fa-globe text-success"
