+++ Products/zms/conf/metaobj_manager/com.zms.foundation/__init__.py
@@ -22,7 +22,7 @@ class com_zms_foundation:
 	package = ""
 
 	# Revision
-	revision = "5.0.0"
+	revision = "5.1.0"
 
 	# Type
 	type = "ZMSPackage"
