+++ Products/zms/plugins/www/themes/themes.zcml
@@ -1,3 +1,3 @@
 <configure xmlns="http://namespaces.zope.org/zope" xmlns:zms="http://namespaces.zope.org/zms">
-  <source location="skin_zms5_base-5.0.1.metaobj.xml"/>
+  <source location="skin_zms5_base-5.0.2.metaobj.xml"/>
 </configure>
