+++ Products/zms/zpt/ZMS/manage_customize.zpt
@@ -402,7 +402,6 @@
 			</form>
 		</div>
 	</div>
-
 </div><!-- #zmi-tab -->
 <footer tal:replace="structure python:here.zmi_body_footer(here,request)">zmi_body_footer</footer>
 </body>
