+++ Products/zms/zpt/ZMSMetacmdProvider/manage_main.zpt
@@ -221,6 +221,11 @@
 			<div class="form-group row mt-5">
 				<div class="controls save">
 					<button name="btn" type="submit" class="btn btn-primary" value="BTN_IMPORT" tal:content="python:here.getZMILangStr('BTN_IMPORT')">Import</button>
+					<a class="btn btn-secondary ml-1" 
+						title="Check-Up: Compare working content model with the current ZMS source files."
+						tal:attributes="href python:'./manage_customize_diff?action=%s&lang=%s&provider_id=metacmd_manager'%(action,request['lang'])" target="_blank">
+						<i class="fas fa-stethoscope text-muted"></i> <span class="badge badge-info">ZMS</span>
+					</a>
 				</div>
 			</div>
 		</form>
