+++ Products/zms/zpt/ZMSMetamodelProvider/manage_main_import.zpt
@@ -42,12 +42,17 @@
 			</div><!-- .form-group -->
 			<div class="badge badge-dark position-absolute mr-1" style="top:0;right:0">Step-1</div>
 			<div class="form-group row">
-				<div class="controls save mt-3">
-					<button type="submit" name="btn" class="btn btn-primary" value="BTN_IMPORT" 
-						tal:content="python:here.getZMILangStr('BTN_IMPORT')">
-						Import
-					</button>
+			<div class="form-group row mt-5">
+				<div class="controls save">
+					<button name="btn" type="submit" class="btn btn-primary" value="BTN_IMPORT" tal:content="python:here.getZMILangStr('BTN_IMPORT')">Import</button>
+					<a class="btn btn-secondary ml-1" 
+						title="Check-Up: Compare working content model with the current ZMS source files."
+						tal:attributes="href python:'./manage_customize_diff?action=%s&lang=%s&provider_id=metaobj_manager'%('manage_changeProperties',request['lang'])" target="_blank">
+						<i class="fas fa-stethoscope text-muted"></i> <span class="badge badge-info">ZMS</span>
+					</a>
 				</div>
+			</div>
+
 			</div><!-- .form-group -->
 		</form>
